mosaic --> grid_utils organization
