+++ libFMS/Makefile.am
@@ -39,7 +39,7 @@ libFMS_la_LIBADD += $(top_builddir)/memutils/libmemutils.la
 libFMS_la_LIBADD += $(top_builddir)/fms/libfms.la
 libFMS_la_LIBADD += $(top_builddir)/fms2_io/libfms2_io.la
 libFMS_la_LIBADD += $(top_builddir)/affinity/libfms_affinity.la
-libFMS_la_LIBADD += $(top_builddir)/mosaic/libmosaic.la
+libFMS_la_LIBADD += $(top_builddir)/grid_utils/libgrid_utils.la
 libFMS_la_LIBADD += $(top_builddir)/mosaic2/libmosaic2.la
 libFMS_la_LIBADD += $(top_builddir)/coupler/libcoupler.la
 libFMS_la_LIBADD += $(top_builddir)/drifters/libdrifters.la
